+++ cogs/handle/errors.py
@@ -1,0 +1,27 @@
+from discord.ext import commands
+from bot import PRODUCTION
+
+
+class Errors(commands.Cog):
+    def __init__(self, bot: commands.Bot):
+        self.bot: commands.Bot = bot
+        self.e: str = "<:ge:767823523573923890>"
+
+    @commands.Cog.listener()
+    async def on_command_error(self, ctx: commands.Context, error):
+        if isinstance(error, commands.MissingRequiredArgument):
+            await ctx.send(f"{self.e}  You didn't pass in all of the arguments, **use** `git --help` **for info.**")
+        elif isinstance(error, commands.CommandOnCooldown):
+            msg = self.e + " " + '**You\'re on cooldown!** Please try again in {:.2f}s'.format(error.retry_after)
+            await ctx.send(msg)
+        elif isinstance(error, commands.MaxConcurrencyReached):
+            await ctx.send(
+                f"{self.e}  This command is experiencing exceptional traffic. **Please try again in a few seconds.**")
+        elif not PRODUCTION:
+            raise error
+        else:
+            print(error)
+
+
+def setup(bot):
+    bot.add_cog(Errors(bot))
